Refactor experiment graphs to use graph-support
